Prefactor types into a separarate types package
This came up in #37 where we ended up in a dependency loop.
- main package was importing a specific rate limiter type
- both rate limiters needed to import Clock types from main package

By moving it to a separate package we avoid that loop.

If we decide to go ahead with #37, this will also be necessary
to have a "common config type" that can be re-used between all N
alternative limiters.
