+++ internal/types/types.go
@@ -1,0 +1,39 @@
+// Copyright (c) 2016, 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package types
+
+import "time"
+
+// Limiter is used to rate-limit some process, possibly across goroutines.
+// The process is expected to call Take() before every iteration, which
+// may block to throttle the goroutine.
+type Limiter interface {
+	// Take should block to make sure that the RPS is met.
+	Take() time.Time
+}
+
+// Clock is the minimum necessary interface to instantiate a rate limiter with
+// a clock or mock clock, compatible with clocks created using
+// github.com/andres-erbsen/clock.
+type Clock interface {
+	Now() time.Time
+	Sleep(time.Duration)
+}
